requirements.txt
tinkoff-investments
python-telegram-bot